[resotolib][chore] Cleanup utils
