+++ resotolib/test/durations_test.py
@@ -5,7 +5,7 @@ from typing import Callable, Any, cast, Optional, TypeVar
 from hypothesis import given
 from hypothesis.strategies import sampled_from, tuples, integers, composite, lists, SearchStrategy, just
 
-from resotolib.durations import time_unit_parser, time_units, parse_duration, DurationRe
+from resotolib.durations import time_unit_parser, time_units, parse_duration, DurationRe, duration_str
 
 units_gen = sampled_from(list(chain.from_iterable(names for _, names, _ in time_units)))
 combines_gen = sampled_from(["", ", ", " and "])
@@ -64,3 +64,17 @@ def test_parse_duration() -> None:
     assert parse_duration("32days, 4hours and 3min and 3s") == timedelta(days=32, hours=4, minutes=3, seconds=3)
     assert parse_duration("-32days, 4hours and 3min and 3s") == timedelta(days=-32, hours=-4, minutes=-3, seconds=-3)
     assert parse_duration("3d4h6m5s") == timedelta(days=3, hours=4, minutes=6, seconds=5)
+
+
+def test_duration_string() -> None:
+    duration = timedelta(days=1, hours=2, minutes=3, seconds=4)
+    assert duration_str(duration) == "1d2h3min4s"
+    # define the most granular unit
+    assert duration_str(duration, down_to_unit="min") == "1d2h3min"
+    assert duration_str(duration, down_to_unit="h") == "1d2h"
+    assert duration_str(duration, down_to_unit="d") == "1d"
+    assert duration_str(duration, down_to_unit="min") == "1d2h3min"
+    # define the precision
+    assert duration_str(duration, precision=1) == "1d"
+    assert duration_str(duration, precision=2) == "1d2h"
+    assert duration_str(duration, precision=3) == "1d2h3min"
